refactor: fix util naming
